Feature/flash messages (#11)
* Fixed some camel case errors

Fixed some missed translations from snake_case to camelCase. Not all fixed yet.

* Fixed last of camelCase bugs

* Removed accidental target files

* Fixed redirect on add message

Instead of redirecting to the user timeline which does not work as no username is given in the request, redirect to the timelime of the logged in user

* Remove accidental target file

* Fixed camelCase in html and mapping

Missed changes from snake_case to camelCase in the timelime.html and the mapping used for render

* Added flash messages

Added flash messages as HTML tags similar to erros. Messages are set using session().attribute() as this was the best suggestion for flash like behaviour I could find

* Made error on login page not include exception

* Fixed wrong user name error message

The error message for entering a wrong username in the login form showed an exception as error message. Has been changed to the intended error message

Co-authored-by: Sif Kristensen <=>