Current progress on CI & CD
Builds maven project and if no errors found deploys it at digital ocean using Github Actions
Database step is not setup yet. Because of this test are not included as of yet.

A docker-compose file has been added for linking the database and the application, but this doesn't work as of now.

Co-authored-by: fremartini <39860858+fremartini@users.noreply.github.com>